move png logic into library
